+++ test/libcell/unit/DiscTypeTest.cpp
@@ -34,10 +34,10 @@ TEST(DiscTypeTest, SettersThrow)
 
     EXPECT_ANY_THROW(dt.setColor(sf::Color::Black));
 
-    for (float i = -1; i <= 0.f; ++i)
+    for (int i = -1; i <= 0; ++i)
     {
-        EXPECT_ANY_THROW(dt.setMass(i));
-        EXPECT_ANY_THROW(dt.setRadius(i));
+        EXPECT_ANY_THROW(dt.setMass(static_cast<float>(i)));
+        EXPECT_ANY_THROW(dt.setRadius(static_cast<float>(i)));
     }
 
     EXPECT_ANY_THROW(dt.setName(""));
